Support endpointslice
