+++ source/Calamari/Kubernetes/ResourceStatus/Resources/EndpointEntry.cs
@@ -1,0 +1,11 @@
+using Newtonsoft.Json;
+
+namespace Calamari.Kubernetes.ResourceStatus.Resources
+{
+    public class EndpointEntry
+    {
+        [JsonProperty("addresses")]
+        public string[] Addresses { get; set; }
+    }
+}
+
