+++ source/Calamari/Kubernetes/ResourceStatus/Resources/EndpointSlice.cs
@@ -6,17 +6,32 @@ namespace Calamari.Kubernetes.ResourceStatus.Resources
 {
     public class EndpointSlice : Resource
     {
+        public string AddressType { get; }
+        public IEnumerable<string> Ports { get; }
         public IEnumerable<string> Endpoints { get; }
 
         public EndpointSlice(JObject json) : base(json)
         {
-            Endpoints = data.SelectTokens("$.endpoints[*].addresses[0]").Values<string>();
+            AddressType = Field("$.addressType");
+            
+            var ports = data.SelectToken("$.spec.ports")
+                ?.ToObject<PortEntry[]>() ?? new PortEntry[] { };
+            Ports = ports.Select(port => port.Port);
+
+            var endpoints = data.SelectToken("$.endpoints")
+                ?.ToObject<EndpointEntry[]>() ?? new EndpointEntry[] { };
+            Endpoints = FormatEndpoints(endpoints);
         }
     
         public override bool HasUpdate(Resource lastStatus)
         {
             var last = CastOrThrow<EndpointSlice>(lastStatus);
-            return !last.Endpoints.SequenceEqual(Endpoints);
+            return last.AddressType != AddressType || !last.Ports.SequenceEqual(Ports) || !last.Endpoints.SequenceEqual(Endpoints);
+        }
+
+        private static IEnumerable<string> FormatEndpoints(IEnumerable<EndpointEntry> endpoints)
+        {
+            return endpoints.SelectMany(endpoint => endpoint.Addresses);
         }
     }
 }
